requirements.txt
Flask >= 3.0.3
Flask-CouchDB >= 0.2.1
python-dotenv >= 1.0.1
flask-cors >= 5.0.0
google-auth >= 2.34.0
requests >= 2.32.3

# Flask psycopg2-binary
# SQLAlchemy